Split GeneralDDLIT to JDBCGeneralDDLIT and ProxyGeneralDDLIT (#8828)
* Rename ActualDataSourceBuilder

* Add ActualDataSourceBuilder.createActualDataSources()

* Refactor BaseIT

* Add JDBCGeneralDDLIT

* Refactor JDBCGeneralDDLIT

* Refactor BaseDDLIT

* Refactor ProxyGeneralDDLIT

* Move IntegrateTestParameters
